fix(competencias): unidade com matriz aparece na avaliacao do gestor
A intersecao no front usava Set de ids vindos de /unidades-com-matriz
(bigint serializado como string) contra os ids de /unidades-lideranca
(int4 serializado como numero), entao Set<string>.has(number) sempre
falhava e o diretor via 'Nenhuma unidade autorizada' mesmo com o
referencial do gestor validado. Coage os dois lados para numero no
front e passa o endpoint a devolver Integer (a coluna e int4).

+++ backend/src/main/java/br/jus/tjgo/kaizen/service/CompetenciasGestorService.java
@@ -400,13 +400,18 @@ public class CompetenciasGestorService {
                 unidadeId);
     }
 
-    /** Ids das unidades com a Matriz de Competências (do tipo) validada até a camada final. */
-    public List<Long> unidadesComMatrizValidada(String tipo) {
+    /**
+     * Ids das unidades com a Matriz de Competências (do tipo) validada até a camada final.
+     * Retorna Integer (a coluna unidade_id é int4) para o JSON sair como número — o front
+     * cruza esses ids com os de /unidades-lideranca, que também vêm como número. Long seria
+     * serializado como string pelo JacksonConfig e quebraria a interseção por tipo.
+     */
+    public List<Integer> unidadesComMatrizValidada(String tipo) {
         return jdbc.queryForList(
                 "SELECT DISTINCT unidade_id FROM competencias_gestor_formularios " +
                         "WHERE tipo = ? AND is_deleted = FALSE AND validado_final_em IS NOT NULL " +
                         "  AND unidade_id IS NOT NULL",
-                Long.class, tipo);
+                Integer.class, tipo);
     }
 
     public List<Map<String, Object>> findCompetenciasGestorByUnidade(long unidadeId) {
